pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "model-forge"
version = "1.0.0"
description = "AI-Powered 3D Model Generation Pipeline: Description → Prompt → Image → 3D Model"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Grid AutoPilot Team"}
]
keywords = ["3D", "model", "generation", "AI", "Gemini", "digital-twin"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "google-genai>=1.0.0",
    "requests>=2.28.0",
    "httpx>=0.23.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
model-forge = "model_forge.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/model-forge"
Documentation = "https://github.com/your-org/model-forge#readme"
Repository = "https://github.com/your-org/model-forge.git"
Issues = "https://github.com/your-org/model-forge/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["model_forge*"]

[tool.setuptools.package-data]
model_forge = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
